+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
      "modules/ismp/pallet/rpc",
      "modules/ismp/pallet/runtime-api",
      "modules/ismp/pallet/demo",
+     "modules/ismp/pallet/testsuite",
      "modules/ismp/testsuite",
      "modules/ismp/clients/sync-committee",
      "modules/ismp/clients/casper-ffg",
