+++ evm/integration-tests/Cargo.toml
@@ -33,13 +33,13 @@ sp-trie = { workspace = true }
 ethers = { workspace = true }
 forge-testsuite = { workspace = true }
 merkle-mountain-range = { workspace = true, features = ["default"] }
-mmr-utils = { path = "../../modules/trees/mmr" }
-ismp = { path = "../../modules/ismp/core" }
-pallet-ismp = { path = "../../modules/ismp/pallet" }
+mmr-utils = { workspace = true }
+ismp = { workspace = true }
+pallet-ismp = { workspace = true }
 ismp-solidity-abi = { path = "../abi", features = ["build-abi"] }
-beefy-prover = { path = "../../modules/consensus/beefy/prover", default-features = false }
-beefy-verifier-primitives = { path = "../../modules/consensus/beefy/primitives" }
-pallet-ismp-relayer = { path = "../../modules/ismp/pallet/relayer" }
+beefy-prover = { workspace = true }
+beefy-verifier-primitives = { workspace = true }
+pallet-ismp-relayer = { workspace = true }
 
 rs_merkle = { git = "https://github.com/polytope-labs/rs-merkle", branch = "seun/2d-merkle-proofs" }
 foundry-evm = { git = "https://github.com/polytope-labs/foundry", rev = "068ef4d82e8421f9f96fdc33b061dc4848b7019a" }
