+++ evm/integration-tests/src/tests/post_request.rs
@@ -13,7 +13,7 @@ use ismp_solidity_abi::{
     handler::{PostRequestLeaf, PostRequestMessage},
     shared_types::StateCommitment,
 };
-use pallet_ismp::mmr_primitives::{DataOrHash, Leaf};
+use pallet_ismp::mmr::primitives::{DataOrHash, Leaf};
 use std::{env, path::PathBuf};
 
 #[tokio::test(flavor = "multi_thread")]
