+++ modules/ismp/pallet/testsuite/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod pallet_ismp;
+mod pallet_ismp_relayer;
