+++ parachain/runtimes/nexus/src/lib.rs
@@ -70,7 +70,7 @@ use frame_system::{
     EnsureRoot, Phase,
 };
 use pallet_ismp::{
-    mmr_primitives::{Leaf, LeafIndex},
+    mmr::primitives::{Leaf, LeafIndex},
     primitives::Proof,
     ProofKeys,
 };
